+++ client/src/js/tasks/tasks_controller.js
@@ -18,9 +18,17 @@ angular.module('trackit.tasks.controller', [
     var categoryId = $stateParams.categoryId;
     $scope.task = {};
 
+    categoriesService.getOne(categoryId).then(function(category) {
+      $scope.category = category;
+    });
+
     tasksService.getAll(categoryId).then(function(tasks) {
       $scope.tasks = tasks;
     });
+
+    $scope.gotoCategories = function () {
+      $state.go('categories', {categoryId: categoryId});
+    };
 
     $scope.createTask = function () {
       tasksService.create({
@@ -33,7 +41,17 @@ angular.module('trackit.tasks.controller', [
       });
     };
 
+    $scope.edit = function () {
+      $state.go('categories.edit', {categoryId: categoryId});
+    };
+
+    $scope.gotoAddTaskState = function () {
+      $state.go('categories.tasks.add', {categoryId: categoryId});
+    };
+
     $scope.destroyCategory = function () {
+      var yes = confirm('are you sure you want to delete this category?');
+      if (!yes) return;
       categoriesService.destroy(categoryId).then(function () {
         $state.go('categories');
       });
